% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_variable_label_template.R
\name{build_variable_label_template}
\alias{build_variable_label_template}
\title{Generate the template list of labels for a set of datasets}
\usage{
build_variable_label_template(
  ...,
  previous_variable_labels = list(),
  output_file = NULL
)
}
\arguments{
\item{previous_variable_labels}{list object of previous labels that should be kept}

\item{output_file}{if given, then the structure will be saved as a yaml for editing in a text editor}
}
\value{
A list of column names for each dataset specified
}
\description{
Generate the template list of labels for a set of datasets
}
