% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_var.R
\name{count_levels_horizontal}
\alias{count_levels_horizontal}
\title{Print the counts of a categorical variable, horizontally}
\usage{
count_levels_horizontal(x, name)
}
\arguments{
\item{x}{Column vector}

\item{name}{Column name}
}
\value{

}
\description{
Print the counts of a categorical variable, horizontally
}
