% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_var.R
\name{count_levels_vertical}
\alias{count_levels_vertical}
\title{Print the counts of categorical column in a vertical table}
\usage{
count_levels_vertical(x, name)
}
\arguments{
\item{x}{Column vector}

\item{name}{Column name}
}
\value{

}
\description{
Print the counts of categorical column in a vertical table
}
