% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_var.R
\name{data_type}
\alias{data_type}
\title{Print the datatype of the column}
\usage{
data_type(x, name)
}
\arguments{
\item{x}{Column vector}

\item{name}{Column name}
}
\value{

}
\description{
Print the datatype of the column
}
