% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_var.R
\name{describe_var}
\alias{describe_var}
\title{Describe a variable in a Rmarkdown document}
\usage{
describe_var(data, name, ...)
}
\arguments{
\item{data}{Dataframe}

\item{name}{Name of the column to describe}

\item{...}{List of display functions to use}
}
\value{
nothing
}
\description{
Describe a variable in a Rmarkdown document
}
