% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_var.R
\name{detect_standard_functions}
\alias{detect_standard_functions}
\title{Infer a set of standard display functions, based on the column provided}
\usage{
detect_standard_functions(c)
}
\arguments{
\item{x}{Column vector}
}
\value{
list of display functions
}
\description{
Infer a set of standard display functions, based on the column provided
}
