% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_var.R
\name{fact}
\alias{fact}
\title{Print the inputed text, with appropriate line breaks afterwards}
\usage{
fact(x)
}
\arguments{
\item{x}{text}
}
\value{

}
\description{
Print the inputed text, with appropriate line breaks afterwards
}
