% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_and_save_dataset.R
\name{label_and_save_dataset}
\alias{label_and_save_dataset}
\title{Add labels to the variables and values in a dataset}
\usage{
label_and_save_dataset(
  df_base,
  data_labels,
  value_labels,
  output_location,
  filename,
  spss_output = F,
  write_output = T
)
}
\arguments{
\item{write_output}{}
}
\value{

}
\description{
Add labels to the variables and values in a dataset
}
