% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_var.R
\name{nice_summary}
\alias{nice_summary}
\title{Print the summary of the values}
\usage{
nice_summary(x, name)
}
\arguments{
\item{x}{Column vector}

\item{name}{Column name}
}
\value{

}
\description{
Print the summary of the values
}
