% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_var.R
\name{show_histogram}
\alias{show_histogram}
\title{Print the histogram of the column}
\usage{
show_histogram(x, name)
}
\arguments{
\item{x}{Column vector}

\item{name}{Column name}
}
\value{

}
\description{
Print the histogram of the column
}
