% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_var.R
\name{summarize_labels}
\alias{summarize_labels}
\title{Print the labels in a summarized form}
\usage{
summarize_labels(x, name)
}
\arguments{
\item{x}{Column vector}

\item{name}{Column name}
}
\value{

}
\description{
Print the labels in a summarized form
}
