% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_var.R
\name{varname}
\alias{varname}
\title{Print the name of the variable}
\usage{
varname(x, name)
}
\arguments{
\item{x}{Column vector}

\item{name}{Column name}
}
\value{

}
\description{
Print the name of the variable
}
